\name{@ANNOBJPREFIX@.db internals}
\alias{miRNAAnnDbBimap}
\alias{miRNAAnnDbBimap-class}
\alias{targetscanTarget}
\alias{targetscanTarget-class}
\alias{miRNATargetAnnDbBimap}
\alias{miRNATargetAnnDbBimap-class}
\alias{targetscanMiRBase}
\alias{targetscanMiRBase-class}
\alias{miRFamily}
\alias{UTRstart}
\alias{UTRend}
\alias{MSAstart}
\alias{MSAend}
\alias{Seedmatch}
\alias{PCT}
\alias{MiRBaseID}
\alias{MiRBaseAccession}
\alias{Seedm8}
\alias{Species}
\alias{Maturesequence}
\alias{Familyconservation}

\title{@ANNOBJPREFIX@.db internals}

\description{
  Objects, classes and methods from the @ANNOBJPREFIX@.db package that are not
  intended to be used directly.
}

\keyword{internal}
\keyword{methods}
\keyword{classes}
